requirements.txt
Django==2.2.2
django-bootstrap-datepicker-plus==3.0.5
pytz==2019.1
sqlparse==0.3.0
